refactor(core): otimizacao N+1, seguranca JWT e precisao no motor de custos
✅ Refatorações aplicadas (Fases 1, 2, 3 e 4):
1. Otimização de Performance (N+1 Queries):
   - Injeção de 'fetch = FetchType.LAZY' nos relacionamentos @ManyToOne das entidades 'Projeto' e 'Timesheet'.
2. Segurança e Information Hiding:
   - Criação do Enum 'Role' e substituição do uso frágil de Strings.
   - Adição do '@JsonIgnore' na senha e atualização da infraestrutura do 'pom.xml' suportando JWT e Spring Security.
3. Precisão Matemática no Core Business:
   - Troca de Integer para Double nas métricas da entidade 'Timesheet' e DTOs, suportando lançamento de horas fracionadas para o Burn Rate.
4. Consistência de Camadas (Services & DTOs):
   - Refatoração das assinaturas em 'TimesheetServiceImpl' e anotações nos pacotes de Request/Response DTOs, blindando os contratos da API.
   - Atualização do 'maven-compiler-plugin' para compatibilidade com Lombok no Java 21+.

+++ backend/src/main/java/com/devflow/dto/TimesheetResponseDto.java
@@ -8,8 +8,8 @@ import lombok.Data;
 public class TimesheetResponseDto {
     private Long id;
     private LocalDate dataRegistro;
-    private Integer horasTrabalhadas;
-    private Integer horasExtras;
+    private Double horasTrabalhadas;
+    private Double horasExtras;
     private String descricaoTarefa;
 
     //Data Flattening (Achatamento)
